Refactor wms and wfs to use org.geoserver.geopkg consistently
