+++ src/community/geopkg/src/main/java/org/geoserver/geopkg/wps/GeoPackageGetMapOutputFormatWPS.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import java.util.SortedMap;
 import java.util.TreeMap;
 import java.util.logging.Level;
+import org.geoserver.geopkg.wms.GeoPackageGetMapOutputFormat;
 import org.geoserver.gwc.GWC;
 import org.geoserver.ows.util.OwsUtils;
 import org.geoserver.platform.GeoServerExtensions;
@@ -16,7 +17,6 @@ import org.geoserver.wms.RasterCleaner;
 import org.geoserver.wms.WMS;
 import org.geoserver.wms.WebMap;
 import org.geoserver.wms.WebMapService;
-import org.geoserver.wms.geopackage.GeoPackageGetMapOutputFormat;
 import org.geotools.data.util.DefaultProgressListener;
 import org.geotools.geometry.jts.ReferencedEnvelope;
 import org.geotools.geopkg.GeoPackage;
