+++ src/extension/geopkg-output/src/test/java/org/geoserver/geopkg/GeoPackageGetMapOutputFormatTest.java
@@ -22,6 +22,7 @@ import javax.imageio.ImageIO;
 import javax.xml.namespace.QName;
 import org.geoserver.config.GeoServer;
 import org.geoserver.data.test.SystemTestData;
+import org.geoserver.geopkg.wms.GeoPackageGetMapOutputFormat;
 import org.geoserver.gwc.GWC;
 import org.geoserver.platform.ServiceException;
 import org.geoserver.wms.GetMapRequest;
@@ -29,7 +30,6 @@ import org.geoserver.wms.WMSInfo;
 import org.geoserver.wms.WMSMapContent;
 import org.geoserver.wms.WMSTestSupport;
 import org.geoserver.wms.WebMap;
-import org.geoserver.wms.geopackage.GeoPackageGetMapOutputFormat;
 import org.geoserver.wms.map.RawMap;
 import org.geotools.geometry.jts.ReferencedEnvelope;
 import org.geotools.geopkg.GeoPackage;
